all issues resolved
