+++ apps/frontend/app/(website)/studio/main/PageBottomPane/utils/panelsData.tsx
@@ -1,20 +1,9 @@
-import type { KnownEngines } from "@codemod-com/utilities";
-import ASTViewer from "@studio/main/ASTViewer";
 import {
-  type ContentViewerVariant,
   type PanelData,
   ResizablePanelsIndices,
 } from "@studio/main/PageBottomPane/utils/types";
 import type { Repeat } from "@studio/types/transformations";
 import * as React from "react";
-
-export const getContent =
-  (type: ContentViewerVariant) => (engine: KnownEngines) =>
-    engine === "jscodeshift" ? (
-      <ASTViewer type={type} />
-    ) : (
-      "The AST View is not yet supported for tsmorph"
-    );
 
 const beforePanel: PanelData = {
   relatedAST: ResizablePanelsIndices.BEFORE_AST,
@@ -22,7 +11,6 @@ const beforePanel: PanelData = {
   snippedIndex: ResizablePanelsIndices.BEFORE_SNIPPET,
   type: "before",
   hasBoundResize: true,
-  content: getContent("before"),
   snippetData: {
     header: (
       <span className="flex items-center justify-center justify-content-center">
@@ -41,7 +29,6 @@ const afterPanel: PanelData = {
   snippedIndex: ResizablePanelsIndices.AFTER_SNIPPET,
   type: "after",
   hasBoundResize: false,
-  content: getContent("after"),
   snippetData: {
     header: "After (Expected)",
     diffEditorWrapper: {
@@ -55,7 +42,6 @@ const outputPanel: PanelData = {
   relatedAST: ResizablePanelsIndices.OUTPUT_AST,
   snippedIndex: ResizablePanelsIndices.OUTPUT_SNIPPET,
   type: "output",
-  content: getContent("output"),
   snippetData: {
     header: (
       <span className="flex items-center justify-center justify-content-center">
